STEP 4: Respond to authors#show with js
Here you have it.  We respond to a javascript
request by looking for the `#zoom-in` element.  We
then change its HTML to be the JavaScript
rendering of an ERB tag that renders HTML.
We have to use `escape_javascript` to make sure
that the ERB-interpolated HTML is transformed to
JS.

This is where this method to Ajax rendering in
Rails feels really _wrong_ to me.  I'll try to
clean it up subsequently, but it feels like we
could just render the parial's content locally.
I'll explore this idea after we get it working.