STEP 5: Final Step
Return an author's name (as HTML).

The advantage, theoretically, of using this
partial is that it could be

1.  shared among other views (DRY)
2.  **AND** not only is it DRY among HTML
templates, it's also going to work with any JS
clients (so the resource's data stay dry between
HTML representations and JS representations)

I see the idea here, but it just feels wrong and
really overcomplicated (to me).